Frontend: add vue-router, split LLM viewer into its own route (phase 6a)
Begin separating the VLA viewer from the existing LLM viewer rather than
conflating them. Add vue-router with two routes sharing one app shell:
  /     -> LLM viewer (cloud LLM serving)   -- behavior unchanged
  /vla  -> VLA viewer (edge robotics)       -- stub for now

- src/views/LlmView.vue: the original App.vue body, moved verbatim (provides +
  Header/LeftPannel/Graph), so the LLM viewer is byte-for-byte unchanged.
- src/App.vue: now just <RouterView/>.
- src/router.js: hash-history routes (deep links work on static hosting).
- src/views/VlaView.vue: placeholder, filled in by phase 6c.
- main.js: install the router.

Existing LLM components (Header, LeftPannel, Config, Graph) and all backend
endpoints are untouched. Verified: `vite build` compiles (1293 modules) and the
dev server serves the app.
