+++ design_docs/vla_roofline_analyzer.md
@@ -99,7 +99,7 @@ Each phase reuses the existing roofline + footprint machinery. New work = op han
 4. **Vision encoder + multi-phase orchestration.** A ViT is a single bidirectional transformer forward, so the encoder reuses the golden-locked `analyze()` via `configs/vit.py` + `model_params/vision_encoders.py` (CLIP-L, DINOv2-L, SigLIP-L presets). Added the `patch_embed` op handler and `roofline_model.evaluate_op` (shared roofline-apply, extracted golden-safe). `vla.py` orchestrates **vision-encode → projector → LLM-prefill**, returns a per-phase latency/OPs breakdown, and a **memory-fit verdict** vs the device's `memory_capacity` plus an optional control-frequency budget check. `analyze_vla_cli.py` + `model_params/vla_models.py` presets (runnable `tinyvla_demo`; `openvla_7b` documents real scale). Golden: 15/15 PASS (added ViT + VLA cases). Action-generation phase deferred to step 5. ✅ DONE
 5. **Action-generation phase + new op handlers.** Added the action phase to `vla.py`: **ar** (autoregressive discrete tokens — iterate LLM decode for horizon×tokens/action; OpenVLA), **flow** (diffusion/flow-matching — a separate action expert run num_steps× over the action chunk; π0/Octo), **parallel** (expert once). Added the `linear_attention` op (SARA-RT, O(L)) and a `_transformer_forward` helper that costs an arbitrary transformer at any q/kv length and attention type (enables the expert + future cross-attention). Exposed chunk/steps/Hz on the CLI. Presets: AR (`tinyvla_demo`, `openvla_7b`), flow (`tinyvla_flow_demo`, `pi0_like`), linear-attn (`tinyvla_flow_linear_demo`). Golden: 17/17 PASS. ✅ DONE
    - *Deferred:* backbone-wide linear attention (analyze() still uses quadratic; linear_attention is wired for the action expert only), and flow expert cross-attention to the prefix KV (currently self-attention over the chunk only).
-6. **Surface in frontend:** new controls + "latency vs control budget / memory-fit" readout.
+6. **Surface in frontend — as a SEPARATE VLA viewer, not conflated with the LLM viewer.** The existing LLM-viewer-web (cloud LLM serving) and the new VLA-viewer-web (edge robotics) are distinct products with different knobs/hardware/headline outputs. Approach (decided): **vue-router, shared shell** — `/` = existing LLM viewer (behavior untouched), `/vla` = new. Additive backend: `get_vla_graph.py` + `POST /get_vla_graph` + `GET /get_vla_avaliable` (VLA presets + edge hardware); `/get_graph` untouched. Frontend: extract current `App.vue` body into `LlmView.vue` (one mechanical move), `App.vue` becomes `<router-view>`. VLA view is a **purpose-built phase dashboard** (latency-share bars per phase, big FITS / control-budget verdicts, per-phase metrics) rather than reusing the G6 DAG — `VlaView/VlaConfig/VlaReadout/VlaHeader.vue` new; `utils.js` reused; roofline chart available for per-phase drill-down. Phasing: (6a) router + LlmView extract, prove `/` identical; (6b) backend VLA endpoints; (6c) VLA dashboard; (6d) polish. [IN PROGRESS]
 7. **Seed configs** for OpenVLA and π0 from their HF configs.
 
 ## 7. Environment notes
