+++ frontend/src/main.js
@@ -1,5 +1,7 @@
 import { createApp } from 'vue'
 import App from './App.vue'
+import { router } from './router.js'
 
-const app=createApp(App)
+const app = createApp(App)
+app.use(router)
 app.mount('#app')
